Don't save the buffers, save the ids only.
This approach is more memory friendly because when thread ids are not
reutilized frequently if does not stores the whole buffer information
in memory until the end of the execution.

The change comes with the extra cost of creating a new buffer for
every thread, but not needing to take a write lock on every thread
construction.