Avoid events entry duplication with lock free code.
The original code was duplicating entries because the if filter is in
the id side and not the name. To avoid this at the thread level it
required to read the map but with the read lock taken.

Out solution solved this by performing the filtering and
initialization only once during the global initialization. After that
the vallues are "frozen", so they can be shared bewen threads in a
static way (passing the vector by reference).

This way we ensure that the ids are alway the same and we don't need
to take any lock during the lazy thread initialization.