Register all supported events preserved the event id.
This ensured that the paraver configs for hardware events are
compatible.
On the other hand it limits the number of supported events, but thats
limited by the perf_event crate any way.
All the events wil appear in the .row, with may be an issue, so, not
sure how good or bad is that.